requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
numpy
safetensors
python-multipart
kornia
einops
huggingface_hub
opencv-python
timm
